+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-03-02-1
+
+Changes:
+
+- Search params
+
 ## v2023-03-01-1
 
 Changes: